+++ tests/int_vulnerable_files.rs
@@ -87,7 +87,6 @@ impl TestContext {
             git_repo_timeout: 1800, // 30 minutes
             output_args: OutputArgs { output: None, format: ReportOutputFormat::Pretty },
             no_dedup: true,
-            ignore_tests: false,
             snippet_length: 256,
             baseline_file: None,
             manage_baseline: false,
@@ -153,7 +152,6 @@ impl TestContext {
             git_repo_timeout: 1800, // 30 minutes
             output_args: OutputArgs { output: None, format: ReportOutputFormat::Pretty },
             no_dedup: true,
-            ignore_tests: false,
             snippet_length: 256,
             baseline_file: None,
             manage_baseline: false,